Added testStompHelper
