+++ src/test/java/com/cgm/infolab/WebSocketDeleteEditPrivateTests.java
@@ -4,6 +4,7 @@ import com.cgm.infolab.db.model.ChatMessageEntity;
 import com.cgm.infolab.db.model.UserEntity;
 import com.cgm.infolab.db.model.Username;
 import com.cgm.infolab.helper.TestDbHelper;
+import com.cgm.infolab.helper.TestStompHelper;
 import com.cgm.infolab.model.ChatMessageDto;
 import com.cgm.infolab.model.WebSocketMessageDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -50,11 +51,9 @@ public class WebSocketDeleteEditPrivateTests {
     @LocalServerPort
     public Integer port;
     @Autowired
-    public TestRestTemplate rest;
-    @Autowired
     public TestDbHelper testDbHelper;
     @Autowired
-    public ObjectMapper objectMapper;
+    public TestStompHelper testStompHelper;
 
     WebSocketStompClient websocket;
 
@@ -65,15 +64,7 @@ public class WebSocketDeleteEditPrivateTests {
     public void setupAll(){
         testDbHelper.clearDbExceptForGeneral();
 
-        websocket =
-                new WebSocketStompClient(
-                        new SockJsClient(
-                                List.of(new WebSocketTransport(new StandardWebSocketClient()))));
-
-        MappingJackson2MessageConverter messageConverter = new MappingJackson2MessageConverter();
-        messageConverter.setObjectMapper(objectMapper);
-
-        websocket.setMessageConverter(messageConverter);
+        websocket = testStompHelper.initWebsocket();
 
         testDbHelper.addUsers(user1, userBanana);
 
@@ -127,7 +118,7 @@ public class WebSocketDeleteEditPrivateTests {
 
     @Test
     void whenDeletingMessageThroughWebsocket_deletionMessageIsReceived_andMessageIsActuallyDeleted() throws ExecutionException, InterruptedException, TimeoutException {
-        StompSession client = getStompSession();
+        StompSession client = testStompHelper.getStompSessionForUser1(websocket, port);
 
         BlockingQueue<WebSocketMessageDto> receivedMessagesSender = new ArrayBlockingQueue<>(2);
         BlockingQueue<WebSocketMessageDto> receivedMessagesDestination = new ArrayBlockingQueue<>(2);
@@ -199,7 +190,7 @@ public class WebSocketDeleteEditPrivateTests {
 
     @Test
     void whenDeletionFails_websocketReturnsNull() throws ExecutionException, InterruptedException, TimeoutException {
-        StompSession client = getStompSession();
+        StompSession client = testStompHelper.getStompSessionForUser1(websocket, port);
 
         BlockingQueue<WebSocketMessageDto> receivedMessagesSender = new ArrayBlockingQueue<>(2);
         BlockingQueue<WebSocketMessageDto> receivedMessagesDestination = new ArrayBlockingQueue<>(2);
@@ -244,7 +235,7 @@ public class WebSocketDeleteEditPrivateTests {
 
     @Test
     void whenSendingInvalidDto_toDeleteMessage_nullIsReturned_messageIsNotDeleted() throws ExecutionException, InterruptedException, TimeoutException {
-        StompSession client = getStompSession();
+        StompSession client = testStompHelper.getStompSessionForUser1(websocket, port);
 
         BlockingQueue<WebSocketMessageDto> receivedMessagesSender = new ArrayBlockingQueue<>(2);
         BlockingQueue<WebSocketMessageDto> receivedMessagesDestination = new ArrayBlockingQueue<>(2);
@@ -299,7 +290,7 @@ public class WebSocketDeleteEditPrivateTests {
 
     @Test
     void whenEditingMessageThroughWebsocket_editMessageIsReceived_andMessageIsActuallyEdited() throws ExecutionException, InterruptedException, TimeoutException {
-        StompSession client = getStompSession();
+        StompSession client = testStompHelper.getStompSessionForUser1(websocket, port);
 
         BlockingQueue<WebSocketMessageDto> receivedMessagesSender = new ArrayBlockingQueue<>(2);
         BlockingQueue<WebSocketMessageDto> receivedMessagesDestination = new ArrayBlockingQueue<>(2);
@@ -337,7 +328,7 @@ public class WebSocketDeleteEditPrivateTests {
 
     @Test
     void whenEditFails_websocketReturnsNull() throws ExecutionException, InterruptedException, TimeoutException {
-        StompSession client = getStompSession();
+        StompSession client = testStompHelper.getStompSessionForUser1(websocket, port);
 
         BlockingQueue<WebSocketMessageDto> receivedMessagesSender = new ArrayBlockingQueue<>(2);
         BlockingQueue<WebSocketMessageDto> receivedMessagesDestination = new ArrayBlockingQueue<>(2);
@@ -382,7 +373,7 @@ public class WebSocketDeleteEditPrivateTests {
 
     @Test
     void whenSendingInvalidDto_toEditMessage_nullIsReturned_messageIsNotEdited() throws ExecutionException, InterruptedException, TimeoutException {
-        StompSession client = getStompSession();
+        StompSession client = testStompHelper.getStompSessionForUser1(websocket, port);
 
         BlockingQueue<WebSocketMessageDto> receivedMessagesSender = new ArrayBlockingQueue<>(2);
         BlockingQueue<WebSocketMessageDto> receivedMessagesDestination = new ArrayBlockingQueue<>(2);
@@ -498,43 +489,5 @@ public class WebSocketDeleteEditPrivateTests {
                     Assertions.assertNull(received.getChat());
                     Assertions.assertNull(received.getDelete());
                 });
-    }
-
-    private StompSession getStompSession() throws InterruptedException, ExecutionException, TimeoutException {
-        String basicAuth = basicAuth(user1.getName().value(), "password1");
-
-        ResponseEntity<MyCsrfToken> csrfResponse = rest.exchange(
-                RequestEntity
-                        .get("/csrf")
-                        .header(HttpHeaders.AUTHORIZATION, basicAuth)
-                        .build(),
-                MyCsrfToken.class);
-
-        StompHeaders stompHeaders = new StompHeaders();
-        MyCsrfToken csrf = csrfResponse.getBody();
-        stompHeaders.add(csrf.headerName(), csrf.token());
-
-        WebSocketHttpHeaders headers = setCookies(csrfResponse);
-        StompSession session = websocket
-                .connectAsync(String.format("http://localhost:%d/chat?access_token=%s", port, encodedAuth("user1", "password1")), headers, stompHeaders, new StompSessionHandlerAdapter() {})
-                .get(1, TimeUnit.SECONDS);
-        return session;
-    }
-
-    private static String basicAuth(String user, String password) {
-        return String.format("Basic %s", encodedAuth(user, password));
-    }
-
-    private static String encodedAuth(String user, String password) {
-        return Base64.encodeBase64String(String.format("%s:%s", user, password).getBytes());
-    }
-
-    private static WebSocketHttpHeaders setCookies(ResponseEntity<MyCsrfToken> csrfResponse) {
-        String cookies = csrfResponse.getHeaders().getFirst(HttpHeaders.SET_COOKIE);
-        assert cookies != null;
-        String sessionId = cookies.split(";")[0];
-        WebSocketHttpHeaders headers = new WebSocketHttpHeaders();
-        headers.add(HttpHeaders.COOKIE, sessionId);
-        return headers;
     }
 }
